feat: restructure op-test and cleanup test folder
