+++ Cargo.toml
@@ -52,6 +52,7 @@ statetest-types = { path = "crates/statetest-types", package = "revm-statetest-t
 context = { path = "crates/context", package = "revm-context", version = "4.1.0", default-features = false }
 context-interface = { path = "crates/context/interface", package = "revm-context-interface", version = "4.1.0", default-features = false }
 handler = { path = "crates/handler", package = "revm-handler", version = "4.1.0", default-features = false }
+op-revm = { path = "crates/op-revm", package = "op-revm", version = "4.0.2", default-features = false }
 
 # alloy 
 alloy-eip2930 = { version = "0.2.1", default-features = false }
