--- crates/op-revm/tests/testdata/fully_migrated_test.json
@@ -1,14 +1,0 @@
-{
-  "result": {
-    "Success": {
-      "gas_refunded": 0,
-      "gas_used": 1000,
-      "logs": [],
-      "output": {
-        "Call": "0x040506"
-      },
-      "reason": "Stop"
-    }
-  },
-  "state": {}
-}
