--- crates/op-revm/tests/testdata/migrated/migration_example.json
@@ -1,14 +1,0 @@
-{
-  "result": {
-    "Success": {
-      "gas_refunded": 0,
-      "gas_used": 500,
-      "logs": [],
-      "output": {
-        "Call": "0x010203"
-      },
-      "reason": "Stop"
-    }
-  },
-  "state": {}
-}
