--- crates/op-revm/tests/testdata/revert_test.json
@@ -1,9 +1,0 @@
-{
-  "result": {
-    "Revert": {
-      "gas_used": 100,
-      "output": "0x"
-    }
-  },
-  "state": {}
-}
